refactor: split large files to satisfy max-file-size lint rule
