+++ src/domains/character/service/character-service.test.ts
@@ -266,46 +266,4 @@ describe("characterService", () => {
 		const result = await characterService.longRest("nonexistent");
 		expect(result).toBeNull();
 	});
-
-	// --- AC Override ---
-
-	it("setAcOverride sets override value", async () => {
-		const created = await characterService.createCharacter(validInput);
-		const updated = await characterService.setAcOverride(created.id, 16);
-		expect(updated).not.toBeNull();
-		expect(updated?.armorClass.override).toBe(16);
-	});
-
-	it("setAcOverride clears override with null", async () => {
-		const created = await characterService.createCharacter(validInput);
-		await characterService.setAcOverride(created.id, 16);
-		const updated = await characterService.setAcOverride(created.id, null);
-		expect(updated).not.toBeNull();
-		expect(updated?.armorClass.override).toBeNull();
-	});
-
-	it("setAcOverride returns null for missing character", async () => {
-		const result = await characterService.setAcOverride("nonexistent", 16);
-		expect(result).toBeNull();
-	});
-
-	it("toggleSavingThrowProficiency adds ability key if not present", async () => {
-		const created = await characterService.createCharacter(validInput);
-		const updated = await characterService.toggleSavingThrowProficiency(created.id, "STR");
-		expect(updated).not.toBeNull();
-		expect(updated?.savingThrowProficiencies).toContain("STR");
-	});
-
-	it("toggleSavingThrowProficiency removes ability key if already present", async () => {
-		const created = await characterService.createCharacter(validInput);
-		await characterService.toggleSavingThrowProficiency(created.id, "STR");
-		const updated = await characterService.toggleSavingThrowProficiency(created.id, "STR");
-		expect(updated).not.toBeNull();
-		expect(updated?.savingThrowProficiencies).not.toContain("STR");
-	});
-
-	it("toggleSavingThrowProficiency returns null for missing character", async () => {
-		const result = await characterService.toggleSavingThrowProficiency("nonexistent", "STR");
-		expect(result).toBeNull();
-	});
 });
