+++ src/domains/character/ui/CharacterSheet.tsx
@@ -1,10 +1,12 @@
 import { useEffect, useState } from "react";
 import { useNavigate } from "../../../app/router.tsx";
-import { calculateAC, getAbilityModifier } from "../types/character.js";
+import { getAbilityModifier } from "../types/character.js";
 import type { Character } from "../types/index.js";
-import { SKILLS, calculateSavingThrow, calculateSkillBonus } from "../types/skills.js";
+import { SKILLS, calculateSkillBonus } from "../types/skills.js";
+import { ArmorClassSection } from "./ArmorClassSection.tsx";
 import styles from "./CharacterSheet.module.css";
 import { EquipmentSection } from "./EquipmentSection.tsx";
+import { SavingThrowsSection } from "./SavingThrowsSection.tsx";
 
 const ABILITY_KEYS = ["STR", "DEX", "CON", "INT", "WIS", "CHA"] as const;
 
@@ -18,9 +20,6 @@ export function CharacterSheet({ id }: { id: string }) {
 	const [character, setCharacter] = useState<Character | null>(null);
 	const [loading, setLoading] = useState(true);
 	const [notes, setNotes] = useState("");
-	const [acOverrideInput, setAcOverrideInput] = useState("");
-	const [showAcOverride, setShowAcOverride] = useState(false);
-
 	useEffect(() => {
 		fetch(`/api/characters/${id}`)
 			.then((r) => (r.ok ? r.json() : null))
@@ -120,49 +119,6 @@ export function CharacterSheet({ id }: { id: string }) {
 			.catch(() => {});
 	};
 
-	const handleSetAcOverride = () => {
-		const value = Number.parseInt(acOverrideInput, 10);
-		if (Number.isNaN(value) || value < 0) return;
-		fetch(`/api/characters/${id}/ac`, {
-			method: "PUT",
-			headers: { "Content-Type": "application/json" },
-			body: JSON.stringify({ override: value }),
-		})
-			.then((r) => (r.ok ? r.json() : null))
-			.then((data) => {
-				if (data) {
-					setCharacter(data);
-					setShowAcOverride(false);
-					setAcOverrideInput("");
-				}
-			})
-			.catch(() => {});
-	};
-
-	const handleClearAcOverride = () => {
-		fetch(`/api/characters/${id}/ac`, {
-			method: "PUT",
-			headers: { "Content-Type": "application/json" },
-			body: JSON.stringify({ override: null }),
-		})
-			.then((r) => (r.ok ? r.json() : null))
-			.then((data) => {
-				if (data) setCharacter(data);
-			})
-			.catch(() => {});
-	};
-
-	const handleToggleSavingThrow = (abilityKey: string) => {
-		fetch(`/api/characters/${id}/saving-throws/${abilityKey}/toggle`, {
-			method: "POST",
-		})
-			.then((r) => (r.ok ? r.json() : null))
-			.then((data) => {
-				if (data) setCharacter(data);
-			})
-			.catch(() => {});
-	};
-
 	const handleDelete = () => {
 		if (!window.confirm(`Delete ${character?.name ?? "this character"}? This cannot be undone.`))
 			return;
@@ -175,9 +131,6 @@ export function CharacterSheet({ id }: { id: string }) {
 
 	if (loading) return <div className={styles.container}>Loading...</div>;
 	if (!character) return <div className={styles.container}>Character not found.</div>;
-
-	const acValue = calculateAC(character.abilityScores.DEX, character.armorClass);
-	const hasAcOverride = character.armorClass.override !== null;
 
 	const hpPercent = character.hp.max > 0 ? (character.hp.current / character.hp.max) * 100 : 0;
 	const hpFillClass =
@@ -234,84 +187,9 @@ export function CharacterSheet({ id }: { id: string }) {
 				</div>
 			</div>
 
-			<div className={styles.section}>
-				<h2 className={styles.sectionTitle}>Armor Class</h2>
-				<div className={styles.acDisplay}>
-					<div className={styles.acShield}>
-						<span className={styles.acValue} data-testid="ac-value">
-							{acValue}
-						</span>
-						<span className={styles.acLabel}>AC</span>
-					</div>
-					{hasAcOverride && <span className={styles.acOverrideIndicator}>Override</span>}
-				</div>
-				<div className={styles.acActions}>
-					{hasAcOverride ? (
-						<button type="button" className={styles.acClearButton} onClick={handleClearAcOverride}>
-							Clear Override
-						</button>
-					) : !showAcOverride ? (
-						<button
-							type="button"
-							className={styles.acOverrideButton}
-							onClick={() => setShowAcOverride(true)}
-						>
-							Override AC
-						</button>
-					) : (
-						<div className={styles.acOverrideForm}>
-							<input
-								type="number"
-								className={styles.acOverrideInput}
-								value={acOverrideInput}
-								onChange={(e) => setAcOverrideInput(e.target.value)}
-								placeholder="AC value"
-								min="0"
-							/>
-							<button type="button" className={styles.acSetButton} onClick={handleSetAcOverride}>
-								Set
-							</button>
-							<button
-								type="button"
-								className={styles.acCancelButton}
-								onClick={() => {
-									setShowAcOverride(false);
-									setAcOverrideInput("");
-								}}
-							>
-								Cancel
-							</button>
-						</div>
-					)}
-				</div>
-			</div>
+			<ArmorClassSection character={character} characterId={id} onUpdate={setCharacter} />
 
-			<div className={styles.section}>
-				<h2 className={styles.sectionTitle}>Saving Throws</h2>
-				<div className={styles.skillsList}>
-					{ABILITY_KEYS.map((key) => {
-						const proficient = character.savingThrowProficiencies?.includes(key) ?? false;
-						const bonus = calculateSavingThrow(
-							character.abilityScores[key],
-							proficient,
-							character.level,
-						);
-						const formatted = bonus >= 0 ? `+${bonus}` : `${bonus}`;
-						return (
-							<label key={key} className={styles.skillRow} data-testid={`saving-throw-${key}`}>
-								<input
-									type="checkbox"
-									checked={proficient}
-									onChange={() => handleToggleSavingThrow(key)}
-									className={styles.skillCheckbox}
-								/>
-								<span className={styles.skillName}>{key}</span>
-								<span className={styles.skillBonus}>{formatted}</span>
-							</label>
-						);
-					})}
-				</div>
-			</div>
+			<SavingThrowsSection character={character} characterId={id} onUpdate={setCharacter} />
 
 			<div className={styles.section}>
 				<h2 className={styles.sectionTitle}>Skills</h2>
